Adapt to Coq 8.7.
Most of the changes are due to Coq 8.7's definition of IZR and its use
to represent literals. The intricacy of the changes ensures that the
library compiles with Coq 8.5, 8.6, and 8.7.

"Require Interval.BigNumsCompat" deals with the removal of BigZ from
standard Coq; this workaround requires Interval 3.3.0 though. Another
possibility would be to require at least Coq 8.6 and use "From Bignums
Require BigZ" instead.